requirements.txt
openai
google-generativeai
faster-whisper
SpeechRecognition
playsound
PyAudio
